made some part of snippets into functions and fixed the code not running when searching for server messages
